Clarify references to Docker images when pushing apps
When Cloud Foundry runs a "Docker app", it bases the app on the filesystem
and metadata in a Docker container image, but it runs the app instances as
Garden containers, not as containers managed by the Docker container engine.

Also replace irrelevant link to `docker start` documentation with direct
references to the directives providing the default start command for an
app based on a Docker image.